guesthouse/backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-filter==24.2
gunicorn==22.0.0
